SERVICE:FIXED

+++ yiyingliService/src/cn/yiyingli/Servlet/AlipayTransNotifyServlet.java
@@ -92,13 +92,14 @@ public class AlipayTransNotifyServlet extends HttpServlet {
 					Order order = orderService.queryByOrderNo(orderNo);
 
 					if (!((alipayNo.equals(order.getAlipayNo()) || alipayNo.equals(order.getTeacher().getAlipay()))
-							&& (Float.valueOf(money) == Float.valueOf(order.getOriginMoney())))) {
+							&& (Float.valueOf(money).floatValue() == Float.valueOf(order.getOriginMoney())
+									.floatValue()))) {
 
 						WarnUtil.sendWarnToCTO("支付订单出错,信息不符合：" + record);
 						LogUtil.error("支付订单出错,信息不符合：" + record, getClass());
 						continue;
 					}
-					if (order.getSalaryState() == OrderService.ORDER_SALARY_STATE_NEED) {
+					if (order.getSalaryState().shortValue() == OrderService.ORDER_SALARY_STATE_NEED) {
 						order.setSalaryState(OrderService.ORDER_SALARY_STATE_DONE);
 						orderService.update(order, false);
 					} else {
